% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_coess.R
\name{annotate_coess}
\alias{annotate_coess}
\title{Combine and annotate co-essential dataframe}
\usage{
annotate_coess(input_ess = NULL, input_inflec = NULL)
}
\arguments{
\item{input_ess}{data frame, A data frame output from \code{coessential_map()}, Default: NULL}

\item{input_inflec}{data frame, A data frame output from get_inflection_points(), Default: NULL}
}
\value{
The input data frame containing co-essential correlation coefficients will be annotated with an additional column
\code{Candidate_gene} to indicate whether the gene is considered to a possible co-essential gene.
}
\description{
Combines and annotates co-essential data frame with inflection points to determine which genes are
likely to be candidates co-essential genes.
}
\details{
Description of output data frame
\itemize{
\item \code{Candidate_gene} - Logical; Whether the coefficient had a p-value < 0.05 and was at or above (positive curve)
or below (negative curve) of the inflection point.
\item See \code{?coessential_map} for details.
}
}
\examples{
gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

load(paste0(
gretta_data_dir,'/sample_22Q2_ARID1A_coessential_result.rda'), 
envir = environment())
load(paste0(
gretta_data_dir,'/sample_22Q2_ARID1A_coessential_inflection.rda'), 
envir = environment())

annotated_df <- annotate_coess(coess_df, coess_inflection_df)

}
