% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_coessential_df.R
\name{annotate_coessential_df}
\alias{annotate_coessential_df}
\title{Combine and annotate co-essential dataframe}
\usage{
annotate_coessential_df(
  input_coessential_df = NULL,
  input_inflection_points = NULL
)
}
\arguments{
\item{input_coessential_df}{data frame, A data frame output from \code{coessential_map()}, Default: NULL}

\item{input_inflection_points}{data frame, A data frame output from get_inflection_points(), Default: NULL}
}
\value{
The input data frame containing co-essential correlation coefficients will be annotated with an additional column
\code{Candidate_gene} to indicate whether the gene is considered to a possible co-essential gene.
}
\description{
Combines and annotates co-essential data frame with inflection points to determine which genes are
likely to be candidates co-essential genes.
}
\details{
Description of output data frame
\itemize{
\item \code{Candidate_gene} - Logical; Whether the coefficient had a p-value < 0.05 and was at or above (positive curve)
or below (negative curve) of the inflection point.
\item See \code{?coessential_map} for details.
}
}
\examples{
\dontrun{

annotated_coessential_df <- annotate_coessential_df(
input_coessential_df = co_ess_res,
input_inflection_points = inflection_points)

}

}
