% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_df.R, R/d.R
\name{annotate_df}
\alias{annotate_df}
\title{Combine and annotate co-essential or co-expression dataframe}
\usage{
annotate_df(
  input_df = NULL,
  input_inflec = NULL,
  top_n = NULL,
  use_inflection = TRUE
)

annotate_df(
  input_df = NULL,
  input_inflec = NULL,
  top_n = NULL,
  use_inflection = TRUE
)
}
\arguments{
\item{input_df}{data frame, A data frame output from \code{coessential_map()}, \code{rna_express()}, and \code{protein_express()}, Default: NULL}

\item{input_inflec}{data frame, A data frame output from get_inflection_points(), Default: NULL}

\item{top_n}{vector, The number of top/bottom candidates to select (eg. if 10 there will be top top
co-essential and 10 anti-essential genes), Default: NULL}

\item{use_inflection}{logical, TRUE to use inflection points as threshold in input_inflec or
FALSE to use top_n as cutoff , Default: TRUE}
}
\value{
The input data frame containing co-essential correlation coefficients will be annotated with an additional column
\code{Candidate_gene} to indicate whether the gene is considered to a possible co-essential gene.

The input data frame containing co-essential correlation coefficients will be annotated with an additional column
\code{Candidate_gene} to indicate whether the gene is considered to a possible co-essential gene.
}
\description{
Combines and annotates co-essential data frame with inflection points to determine which genes are
likely to be candidates co-essential genes.

Combines and annotates co-essential data frame with inflection points to determine which genes are
likely to be candidates co-essential genes.
}
\details{
Description of output data frame
\itemize{
\item \code{Candidate_gene} - Logical; Whether the coefficient had a p-value < 0.05 and was at or above (positive curve)
or below (negative curve) of the inflection point.
\item See \code{?coessential_map} for details.
}

Description of output data frame
\itemize{
\item \code{Candidate_gene} - Logical; Whether the coefficient had a p-value < 0.05 and was at or above (positive curve)
or below (negative curve) of the inflection point.
\item See \code{?coessential_map} for details.
}
}
\examples{
gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

if(!dir.exists(gretta_data_dir)){
  download_example_data(".")
}

load(paste0(
gretta_data_dir,'/sample_22Q2_ARID1A_coessential_result.rda'), 
envir = environment())
load(paste0(
gretta_data_dir,'/sample_22Q2_ARID1A_coessential_inflection.rda'), 
envir = environment())

annotated_df <- annotate_df(coess_df, coess_inflection_df)

gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

if(!dir.exists(gretta_data_dir)){
  download_example_data(".")
}

load(paste0(
gretta_data_dir,'/sample_22Q2_ARID1A_coessential_result.rda'), 
envir = environment())
load(paste0(
gretta_data_dir,'/sample_22Q2_ARID1A_coessential_inflection.rda'), 
envir = environment())

annotated_df <- annotate_df(coess_df, coess_inflection_df)

}
