% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Append_perm_pval.R
\name{Append_perm_pval}
\alias{Append_perm_pval}
\title{Appends permutation adjusted p-value}
\usage{
Append_perm_pval(GI_screen_res = NULL, perms = NULL)
}
\arguments{
\item{p_value}{dataframe, Output from GI_screen(), Default: NULL}

\item{perm_pvalues}{dataframe, Output from GI_screen_perms(), Default: NULL}
}
\value{
A vector of adjusted p-values
}
\description{
Randomly samples dependency probabilities from a given mutant and control group screened.
}
\details{
P-values are adjusted based on permuted random sampled background data.
}
\examples{
# Please see tutorial

}
