% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLE_exp.R
\docType{data}
\name{CCLE_exp}
\alias{CCLE_exp}
\title{DepMap 20Q1: Cancer cell line RNA-seq expression}
\format{
A data frame with 1279 rows and 19145 variables:
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
CCLE_exp
}
\description{
A data set containing the RNA-seq expression in transcripts per million (TPM) of 1279 cancer cell lines.
Data was generated by the Cancer Cell Line Encyclopedia (CCLE) and distributed by DepMap.
Details on data generation can be found in Ghandi, M., et al (2019) (https://www.nature.com/articles/s41586-019-1186-3).
The data was obtained from 'CCLE_expression.csv' (see source url).
}
\keyword{datasets}
