% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLE_exp_annot.R
\docType{data}
\name{CCLE_exp_annot}
\alias{CCLE_exp_annot}
\title{DepMap 20Q1: Cancer cell line RNA-seq expression annotation}
\format{
A data frame with 19145 rows and 4 variables:
\describe{
\item{\code{names}}{character "CCLE_exp" column names}
\item{\code{GeneNames}}{character Hugo symbols}
\item{\code{GeneID}}{character NCBI gene IDs}
\item{\code{GeneNameID}}{character Hugo symbol separated by NCBI gene ID}
}
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
CCLE_exp_annot
}
\description{
A data set containing gene name column headers from the "CCLE_exp" data frame and
its various forms that exist in the DepMap data set platform.
}
\keyword{datasets}
