% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_coefs_coess.R
\name{common_coefs_coess}
\alias{common_coefs_coess}
\title{Perform Pearson coefficient mapping of all pairs of genes in a essentiality screen}
\usage{
common_coefs_coess(
  input_genes = NULL,
  input_disease = NULL,
  input_cell_lines = NULL,
  core_num = NULL,
  output_dir = NULL,
  data_dir = NULL,
  filename = NULL,
  test = FALSE
)
}
\arguments{
\item{input_genes}{string, A vector containing one or more "Hugo Symbol_NCBIID", Default: NULL}

\item{input_disease}{string, A vector one or more disease contexts, Will perform pan-cancer analyses
(all cell lines) by default, Default: NULL}

\item{input_cell_lines}{string, A vector DepMap_IDs for which co-essentiality mapping will be performed on.
Will perform pan-cancer analyses (all cell lines) by default, Default: NULL}

\item{core_num}{integer, Number of cores to run analysis, Default: NULL}

\item{output_dir}{string, Full path to where output file should be saved, Default: NULL}

\item{data_dir}{string Path to GRETTA_data}

\item{filename}{string name of file without the '.csv' extension.}

\item{test}{logical, TRUE/FALSE whether you want to run only a small subset (first 10 genes) to ensure function will run properly
prior to running all genes. Default: FALSE.}
}
\value{
A data frame containing Pearson correlation coefficients. A copy is also saved to the
directory defined in \code{output_dir}.
}
\description{
Performs Pearson correlation coefficient analyses between all gene pairs found in an essentiality screen. Input parameters should match those
used to run \code{annotate_df()}.
}
\details{
Description of output data frame
\itemize{
\item \code{GeneNameID_A} - Hugo symbol with NCBI gene ID of query gene.
\item \code{GeneNameID_B} - Hugo symbol with NCBI gene ID of all genes targeted in the DepMap KO screen.
\item \code{estimate} - Correlation coefficient output from \code{?cor.test}.
\item \code{statistic} - Pearson's correlation statistic. Output from \code{?cor.test}.
\item \code{p.value} - P-value from Pearson's correlation statistic. Output from \code{?cor.test}.
\item \code{parameter} - Degrees of freedom. Output from \code{?cor.test}.
}
}
\examples{
gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

if(!dir.exists(gretta_data_dir)){
  download_example_data(".")
}

\dontrun{
coess_df <- common_coefs_coess(
input_gene = c("ARID1A", "SMARCB1"),
input_disease = 'Pancreatic Cancer',
core_num = 2,
data_dir = gretta_data_dir, 
output_dir = gretta_output_dir,
test = TRUE)
}

}
