% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_num.R
\docType{data}
\name{copy_num}
\alias{copy_num}
\title{DepMap 20Q1: Cancer cell line copy number}
\format{
A data frame with 1713 rows and 27640 variables:
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
copy_num
}
\description{
A data set containing the copy number of 27640 genes in 1713 cancer cell lines.
Data was generated by the Cancer Cell Line Encyclopedia (CCLE) and distributed by DepMap.
Details on data generation can be found in Ghandi, M., et al (2019) (https://www.nature.com/articles/s41586-019-1186-3).
The data was obtained from 'CCLE_expression.csv' (see source url).
}
\keyword{datasets}
