% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep.R
\docType{data}
\name{dep}
\alias{dep}
\title{DepMap 20Q1: Cancer cell line gene dependency probabilities}
\format{
A data frame with 19145 rows and 4 variables:
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
dep
}
\description{
A data set containing the gene dependency probabilities of 18334 genes in 739 cancer cell lines.
Details on data generation can be found in Meyers, RM., et al (2017).
The data was obtained from 'Achilles_gene_dependency.csv' (see source url).
}
\keyword{datasets}
