% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_annot.R
\docType{data}
\name{dep_annot}
\alias{dep_annot}
\title{DepMap 20Q1: Cancer cell line gene dependency probabilities annotation}
\format{
A data frame with 18334 rows and 4 variables:
\describe{
\item{\code{names}}{character "dep" column names}
\item{\code{GeneNames}}{character Hugo symbols}
\item{\code{GeneID}}{character NCBI gene IDs}
\item{\code{GeneNameID}}{character Hugo symbol separated by NCBI gene ID}
}
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
dep_annot
}
\description{
A data set containing gene name column headers from the "dep" data frame and
its various forms that exist in the DepMap data set platform.
}
\keyword{datasets}
