% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential_genes.R
\docType{data}
\name{essential_genes}
\alias{essential_genes}
\title{DepMap 20Q1: Gold standard essential genes list}
\format{
A data frame with 2290 rows and 3 variables:
\describe{
\item{\code{GeneNameID}}{character character Hugo symbol separated by NCBI gene ID}
\item{\code{GeneNames}}{character Hugo symbols}
\item{\code{GeneID}}{character NCBI gene IDs}
}
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
essential_genes
}
\description{
A data set containing gold standard essential genes as described in Meyers, RM., et al (2017).
The data was obtained from 'common_essentials.csv' (see source url).
}
\keyword{datasets}
