% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_protein_expr.R
\name{extract_protein_expr}
\alias{extract_protein_expr}
\title{Extract protein expression data for given cell lines}
\usage{
extract_protein_expr(Input_samples = NULL, Input_genes = NULL, data_dir = NULL)
}
\arguments{
\item{Input_samples}{string A vector of DepMap_ID(s) must be provided, Default: NULL}

\item{Input_genes}{string Optional Hugo Symbol(s) encoding proteins of interest, Default: NULL}

\item{data_dir}{string Path to GINIR_data}
}
\value{
Data frame containing protein expression for samples provided in the input.
If no genes were specified, the function will return a data frame of proteins profiled in DepMap
}
\description{
FUNCTION_DESCRIPTION
}
\details{
See also \code{extract_protein_rna} to extract proteomics profile data
}
\examples{
\dontrun{
extract_protein_expr(Input_samples = c("ACH-000004", "ACH-000146"), Input_genes = c("ATM","TOP1"))
}
}
