% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rna.R
\name{extract_rna}
\alias{extract_rna}
\title{Extract RNA expression data for given cell lines}
\usage{
extract_rna(input_samples = NULL, input_genes = NULL, data_dir = NULL)
}
\arguments{
\item{input_samples}{string A vector of DepMap_ID(s) must be provided, Default: NULL}

\item{input_genes}{string Optional Hugo Symbol(s), Default: NULL}

\item{data_dir}{string Path to GINIR_data}
}
\value{
Data frame containing RNA expression (TPM) for sample provided in the input.
If no genes were specified, the function will return a data frame of all genes profiled in DepMap
}
\description{
FUNCTION_DESCRIPTION
}
\details{
See also \code{extract_prot} to extract proteomics profile data
}
\examples{
gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

extract_rna(
input_samples = c('ACH-001642','ACH-000688'), 
input_genes = c('ARID1A'),
data_dir = gretta_data_dir)

}
