% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_effect.R
\docType{data}
\name{gene_effect}
\alias{gene_effect}
\title{DepMap 20Q1: Cancer cell line gene KO effects}
\format{
A data frame with 739 rows and 18334 variables:
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
gene_effect
}
\description{
A data set containing the gene KO effect of 18334 genes in 739 cancer cell lines.
Details on data generation can be found in Meyers, RM., et al (2017).
The data was obtained from 'Achilles_gene_effect.csv' (see source url).
}
\keyword{datasets}
