% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inflection_points.R
\name{get_inflection_points}
\alias{get_inflection_points}
\title{Calculate inflection points of co-essential data frame}
\usage{
get_inflection_points(Input_coessential_df = NULL)
}
\arguments{
\item{Input_coessential_df}{data frame, A data frame output from \code{coessential_map()}, Default: NULL.}
}
\value{
A data frame containing rank at which the threshold should be drawn for positive and negative co-essential genes.
}
\description{
Calculates the inflection point of the positive and negative curves to determine threshold for
co-essential genes. See \code{?coessential_map} for details.
}
\details{
Description of output data frame
\itemize{
\item \code{Inflection_point_pos_byRank} - Rank threshold for positive curve.
\item \code{Inflection_point_neg_byRank} - Rank threshold for negative curve.
}
}
\examples{
\dontrun{

inflection_points <- get_inflection_points(Input_coessential_df = coess_df)

}

}
