% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GI_screen.R
\name{GI_screen}
\alias{GI_screen}
\title{Perform genetic interaction screen}
\usage{
GI_screen(
  control_id = NULL,
  mutant_id = NULL,
  gene_list = NULL,
  core_num = NULL,
  output_dir = NULL,
  data_dir = NULL,
  filename = NULL,
  test = FALSE
)
}
\arguments{
\item{control_id}{string, A vector containing two or more DepMap_id, Default: NULL}

\item{mutant_id}{string, A vector containing two or more DepMap_id, Default: NULL}

\item{gene_list}{string, A vector containing a list of Hugo symbols to subset the screen and to perform a small in-silico screen, Default: NULL}

\item{core_num}{integer, Number of cores to run analysis, Default: NULL}

\item{output_dir}{string, Full path to where output file should be saved, Default: NULL}

\item{data_dir}{string Path to GRETTA_data}

\item{filename}{string name of file without the '.csv' extension.}

\item{test}{logical, For test_that to shorten computational time for testing}
}
\value{
A data frame containing results from the genetic screen. A copy is also saved to the
directory defined in \code{output_dir}.
}
\description{
Compares dependency probabilities of mutant and control groups to determine
whether the mutant group are resistant or sensitive to specific gene perturbations.
}
\details{
Description of output data frame
\itemize{
\item \code{GeneName_ID} - Hugo symbol with NCBI gene ID
\item \code{GeneNames} - Hugo symbol
\item \verb{_median}, \verb{_mean}, \verb{_sd}, \verb{_iqr} - Control and mutant group's median, mean, standard deviation (sd),
and interquartile range (iqr) of dependency probabilities. Dependency probabilities range from zero to one,
where one indicates a essential gene (ie. KO of gene was lethal) and zero indicates a non-essential gene
(KO of gene was not lethal)
\item \code{Pval} - P-value from Mann Whitney U test between control and mutant groups.
\item \code{Adj_pval} - BH-adjusted P-value.
\item \code{log2FC_by_median} - Log2 normalized median fold change of dependency probabilities (mutant / control).
\item \code{log2FC_by_mean} - Log2 normalized mean fold change of dependency probabilities (mutant / control).
\item \code{CliffDelta} - Cliff's delta non-parametric effect size between mutant and control dependency probabilities.
Ranges between -1 to 1.
\item \code{dip_pval} - Hartigan's dip test p-value. Tests whether distribution of mutant dependency probability is unimodel.
If dip test is rejected (p-value < 0.05), this indicates that there is a multimodel dependency probability distribution and
that there may be another factor contributing to this separation.
\item \code{Interaction_score} - Combined value generated from signed p-values: \verb{-log10(Pval) \\* sign(log2FC_by_median)}
}
}
\examples{
gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

if(!dir.exists(gretta_data_dir)){
  download_example_data(".")
}

if(!dir.exists(gretta_data_dir)){
  download_example_data(".")
}

Screen_results <- GI_screen(
control_id = c('ACH-001354', 'ACH-000274', 'ACH-001799'), 
mutant_id = c('ACH-000911', 'ACH-001957', 'ACH-000075'), 
gene_list = c('ARID1A', 'ARID1B', 'SMARCA2'),
core_num = 2, 
output_dir = gretta_output_dir,
data_dir = gretta_data_dir)

}
