% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GI_screen_perms.R
\name{GI_screen_perms}
\alias{GI_screen_perms}
\title{Perform permutation test n genetic interaction screen}
\usage{
GI_screen_perms(
  control_id = NULL,
  mutant_id = NULL,
  n_perm = 100,
  core_num = NULL,
  output_dir = NULL,
  data_dir = NULL,
  filename = NULL
)
}
\arguments{
\item{control_id}{string, A vector containing two or more DepMap_id, Default: NULL}

\item{mutant_id}{string, A vector containing two or more DepMap_id, Default: NULL}

\item{n_perm}{integer, Number of permutations to run, Default: 100}

\item{core_num}{integer, Number of cores to run analysis, Default: NULL}

\item{output_dir}{string, Full path to where output file should be saved, Default: NULL}

\item{data_dir}{string Path to GRETTA_data}

\item{filename}{string name of file without the '.csv' extension.}
}
\value{
A data frame containing results from the genetic screen. A copy is also saved to the
directory defined in \code{output_dir}.
}
\description{
Randomly samples dependency probabilities from a given mutant and control group screened.
}
\details{
Description of output data frame
\itemize{
\item \code{perm_test} - permutation iteration
\item \verb{_median}, \verb{_mean} - Control and mutant group's median, mean, of dependency probabilities. Dependency probabilities range from zero to one,
where one indicates a essential gene (ie. KO of gene was lethal) and zero indicates a non-essential gene
(KO of gene was not lethal)
\item \code{Pval} - P-value from Mann Whitney U test between control and mutant groups.
\item \code{n_perm} - Number of permutations assigned
}
}
\examples{
gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

if(!dir.exists(gretta_data_dir)){
  download_example_data(".")
}

if(!dir.exists(gretta_data_dir)){
  download_example_data(".")
}

Screen_results <- GI_screen(
control_id = c('ACH-001354', 'ACH-000274', 'ACH-001799'), 
mutant_id = c('ACH-000911', 'ACH-001957', 'ACH-000075'), 
n_perm = 10,
core_num = 2, 
output_dir = gretta_output_dir,
data_dir = gretta_data_dir)

}
