% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GINI_screen.R
\name{GINI_screen}
\alias{GINI_screen}
\title{Perform genetic interaction screen}
\usage{
GINI_screen(
  control_IDs = NULL,
  mutant_IDs = NULL,
  core_num = NULL,
  output_dir = NULL
)
}
\arguments{
\item{control_IDs}{string, A vector containing two or more DepMap_IDs, Default: NULL}

\item{mutant_IDs}{string, A vector containing two or more DepMap_IDs, Default: NULL}

\item{core_num}{integer, Number of cores to run analysis, Default: NULL}

\item{output_dir}{string, Full path to where output file should be saved, Default: NULL}
}
\value{
A data frame containing results from the genetic screen. A copy is also saved to the
directory defined in \code{output_dir}.
}
\description{
Compares dependency probabilities of mutant and control groups to determine
whether the mutant group are resistant or suceptable specific gene perturbations.
}
\details{
Description of output data frame
\itemize{
\item \code{GeneName_ID} - Hugo symbol with NCBI gene ID
\item \code{GeneNames} - Hugo symbol
\item \verb{_median}, \verb{_mean}, \verb{_sd}, \verb{_iqr} - Control and mutant group's median, mean, standard deviation (sd),
and interquartile range (iqr) of dependency probabilities. Dependency probabilities range from zero to one,
where one indicates a essential gene (ie. KO of gene was lethal) and zero indicates a non-essential gene
(KO of gene was not lethal)
\item \code{Pval} - P-value from Mann Whitney U test between control and mutant groups.
\item \code{log2FC_by_median} - Log2 normalized median fold change of dependency probabilities (mutant / control).
\item \code{log2FC_by_mean} - Log2 normalized mean fold change of dependency probabilities (mutant / control).
\item \code{CliffDelta} - Cliff's delta non-parametric effect size between mutant and control dependency probabilities.
Ranges between -1 to 1.
\item \code{dip_pval} - Hartigan's dip test p-value. Tests whether distribution of mutant dependency probability is unimodel.
If dip test is rejected (p-value < 0.05), this indicates that there is a multimodel dependency probability distribution and
that there may be another factor contributing to this separation.
}
}
\examples{
\dontrun{

Screen_results <- GINI_screen(
control_IDs = c("ACH-001354", "ACH-000274", "ACH-001799"), 
mutant_IDs = c("ACH-000911", "ACH-001957", "ACH-000075"), 
core_num = 2, 
output_dir = "~/Desktop/GINI_test_dir/")

}
}
