% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_cancer_type.R
\name{list_available_cancer_types}
\alias{list_available_cancer_types}
\alias{list_available_cancer_subtypes}
\title{List cancer types that are available}
\usage{
list_available_cancer_types(data_dir)

list_available_cancer_subtypes(input_disease, data_dir)
}
\arguments{
\item{data_dir}{string Path to GINIR_data}

\item{input_disease}{string A vector of unique with one or more cancer types listed in \code{list_available_cancer_types()}}
}
\value{
string A vector containing unique cancer types available
}
\description{
\code{list_available_cancer_types()} and \code{list_available_cancer_subtypes()} provide tools for identifying cancer (sub)types that are available in DepMap.
}
\section{Functions}{
\itemize{
\item \code{list_available_cancer_subtypes()}: List cancer subtypes that are available

}}
\examples{
\dontrun{
list_available_cancer_types(data_dir = "/path/to/data")
}

\dontrun{
list_available_cancer_subtypes("Lung Cancer", data_dir = "/path/to/DepMap_data/")
} 
}
