% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_mutations.R
\name{list_mutations}
\alias{list_mutations}
\title{Search mutations detected in DepMap cancer cell lines}
\usage{
list_mutations(
  gene = NULL,
  chr = NULL,
  start_bp = NULL,
  end_bp = NULL,
  is_hotspot = NULL,
  is_damaging = NULL,
  variant_classification = NULL,
  data_dir = NULL
)
}
\arguments{
\item{gene}{string, Hugo Symbol, Default: NULL}

\item{chr}{string, Default: NULL}

\item{start_bp}{integer, Default: NULL}

\item{end_bp}{integer, Default: NULL}

\item{is_hotspot}{logical, TCGA or COSMIC hotspot, Default: NULL}

\item{is_damaging}{logical, Default: NULL}

\item{variant_classification}{string, Default: NULL.

Select from the following:
For DepMap data version prior to 23Q:
3'UTR, 5'Flank, 5'UTR,
De_novo_Start_OutOfFrame, Frame_Shift_Del, Frame_Shift_Ins, IGR, In_Frame_Del, In_Frame_Ins,
Intron, Missense_Mutation, Nonsense_Mutation, Nonstop_Mutation, Silent, Splice_Site,
Start_Codon_Del, Start_Codon_Ins, Start_Codon_SNP, Stop_Codon_Del, Stop_Codon_Ins

For DepMap data version 23Q and after:
missense_variant, frameshift_variant, splice_acceptor_variant, stop_gained, splice_donor_variant,
inframe_deletion, inframe_insertion, stop_lost, start_lost, downstream_gene_variant,
protein_altering_variant, 5_prime_UTR_variant, upstream_gene_variant, non_coding_transcript_exon_variant}

\item{data_dir}{string Path to GRETTA_data}
}
\value{
A data frame containing mutations matching criteria of input arguments
}
\description{
Quick method to look up mutations of interest prior to using \code{select_cell_lines()}
}
\examples{
gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

if(!dir.exists(gretta_data_dir)){
  download_example_data(".")
}

list_mutations(
gene = 'ARID1A', 
is_damaging = TRUE,
data_dir = gretta_data_dir)

}
