% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonessential_genes.R
\docType{data}
\name{nonessential_genes}
\alias{nonessential_genes}
\title{DepMap 20Q1: Common non-essential genes list}
\format{
A data frame with 2290 rows and 3 variables:
\describe{
\item{\code{GeneNameID}}{character character Hugo symbol separated by NCBI gene ID}
\item{\code{GeneNames}}{character Hugo symbols}
\item{\code{GeneID}}{character NCBI gene IDs}
}
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
nonessential_genes
}
\description{
A data set containing common non-essential genes as described in Meyers, RM., et al (2017).
The data was obtained from 'nonessentials.csv' (see source url).
}
\keyword{datasets}
