% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coess.R
\name{plot_coess}
\alias{plot_coess}
\title{Plot coessential genes}
\usage{
plot_coess(
  result_df = NULL,
  inflection_df = NULL,
  label_genes = FALSE,
  label_n = NULL
)
}
\arguments{
\item{result_df}{data frame, A data frame output from \code{annotate_coessential_df()}, Default: NULL}

\item{inflection_df}{data frame, A data frame output from \code{get_inflection_points()}, Default: NULL}

\item{label_genes}{logical, TRUE to trigger gene name labeling, Default: FALSE}

\item{label_n}{integer, Number of genes from either end to label, Default: 1}
}
\value{
A plot generated by ggplot2, additional ggplot layers can be applied directly using \code{+}
}
\description{
Generates a ranked scatter plot of cossential genes from \code{annotate_coessential_df} outputs
}
\examples{
gretta_data_dir <- '/projects/marralab/ytakemon_prj/DepMap/GRETTA_data/22Q2/data'
gretta_output_dir <- '/projects/marralab/ytakemon_prj/DepMap/GRETTA_troubleshooting/'

load(paste0(
gretta_data_dir,'/sample_22Q2_ARID1A_coessential_result.rda'), 
envir = environment())
load(paste0(
gretta_data_dir,'/sample_22Q2_ARID1A_coessential_inflection.rda'), 
envir = environment())

plot_coess(
result_df = coess_df, 
inflection_df = coess_inflection_df, 
label_genes = FALSE)

}
