% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coessential_genes.R
\name{plot_coessential_genes}
\alias{plot_coessential_genes}
\title{Plot coessential genes}
\usage{
plot_coessential_genes(
  result_df = NULL,
  inflection_df = NULL,
  label_genes = FALSE,
  label_n = NULL
)
}
\arguments{
\item{result_df}{data frame, A data frame output from \code{annotate_coessential_df()}, Default: NULL}

\item{inflection_df}{data frame, A data frame output from \code{get_inflection_points()}, Default: NULL}

\item{label_genes}{logical, TRUE to trigger gene name labeling, Default: FALSE}

\item{label_n}{integer, Number of genes from either end to label, Default: 1}
}
\value{
A plot generated by ggplot2, additional ggplot layers can be applied directly using \code{+}
}
\description{
Generates a ranked scatter plot of cossential genes from \code{annotate_coessential_df} outputs
}
\examples{
\dontrun{
plot_coessential_genes(
result_df = coess_df, 
inflection_df = inflection_points_df, 
label_genes = TRUE, 
label_n = 3)
}

}
