% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_screen.R
\name{plot_screen}
\alias{plot_screen}
\title{Plot genetic screen results}
\usage{
plot_screen(result_df = NULL, label_genes = FALSE, label_n = NULL)
}
\arguments{
\item{result_df}{data frame, A data frame output from \code{GINI_screen()}, Default: NULL}

\item{label_genes}{logical, TRUE to trigger gene name labeling, Default: FALSE}

\item{label_n}{integer, Number of genes from either end to label, Default: 1}
}
\value{
A plot generated by ggplot2, additional ggplot layers can be applied directly using \code{+}
}
\description{
Generates a ranked scatter plot of genetic interaction scores from \code{GINI_screen} outputs
}
\examples{
gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

load(paste0(gretta_data_dir,"/sample_22Q2_ARID1A_KO_screen.rda"), envir = environment())

plot_screen(screen_results, label_genes = FALSE)

}
