% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_annot.R
\docType{data}
\name{protein_annot}
\alias{protein_annot}
\title{DepMap 20Q1: Cancer cell line protein expression annotation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 426 rows and 2 columns.
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
protein_annot
}
\description{
A data set containing cell line column headers from the "protein" data frame and
its various forms that exist in the DepMap data set platform.

@format A data frame with 426 rows and 2 variables:
\describe{
\item{\code{GygiNames}}{character "protein" data frame cell line column names}
\item{\code{DepMap_ID}}{character Corresponding DepMap_ID}
}
}
\keyword{datasets}
