% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_nodup.R
\docType{data}
\name{protein_nodup}
\alias{protein_nodup}
\title{DepMap 20Q1: Cancer cell line protein expression with only the major isoform}
\format{
A data frame with 4893 rows and 426 variables:
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
protein_nodup
}
\description{
A data set containing the major protein isoform expression of 426 cancer cell lines.
Details on data generation can be found in Nusinow, DP., et al (2020) (https://www.sciencedirect.com/science/article/pii/S0092867419313856).
The data was obtained from 'protein_quant_current_normalized.csv' (see source url).
}
\keyword{datasets}
