% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pval_perm_adjust.R
\name{Pval_perm_adjust}
\alias{Pval_perm_adjust}
\title{Perform permutation test n genetic interaction screen}
\usage{
Pval_perm_adjust(p_value = NULL, perm_pvalues = NULL)
}
\arguments{
\item{p_value}{numeric, A vector of p-value generated by GI_screen(), Default: NULL}

\item{perm_pvalues}{numeric, A vector of p-values generated by GI_screen_perms(), Default: NULL}
}
\value{
A vector of adjusted p-values
}
\description{
Randomly samples dependency probabilities from a given mutant and control group screened.
}
\details{
P-values are adjusted based on permuted random sampled background data.
}
\examples{
# Please see tutorial

}
