% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_annot.R
\docType{data}
\name{sample_annot}
\alias{sample_annot}
\title{DepMap 20Q1: Cancer cell line sample annotations}
\format{
A data frame with 1775 rows and 24 variables:
\describe{
\item{\code{DepMap_ID}}{DepMap designated IDs used as a use key}
\item{\code{stripped_cell_line_name}}{character}
\item{\code{CCLE_Name}}{character}
\item{\code{alias}}{character}
\item{\code{COSMIC_ID}}{double}
\item{\code{lineage}}{character}
\item{\code{lineage_subtype}}{character}
\item{\code{lineage_sub_subtype}}{character}
\item{\code{lineage_molecular_subtype}}{character}
\item{\code{sex}}{character}
\item{\code{source}}{character}
\item{\code{Achilles_n_replicates}}{double}
\item{\code{cell_line_NNMD}}{double}
\item{\code{culture_type}}{character}
\item{\code{culture_medium}}{character}
\item{\code{cas9_activity}}{character}
\item{\code{RRID}}{character}
\item{\code{sample_collection_site}}{character}
\item{\code{primary_or_metastasis}}{character}
\item{\code{disease}}{character}
\item{\code{disease_subtype}}{character}
\item{\code{age}}{double}
\item{\code{Sanger_model_ID}}{character}
\item{\code{additional_info}}{character}
}
}
\source{
\url{https://figshare.com/articles/dataset/DepMap_20Q1_Public/11791698}
}
\usage{
sample_annot
}
\description{
A data set containing the cancer cell lines information generated by the Cancer Cell Line Encyclopedia (CCLE) and distributed by DepMap.
The data was obtained from 'sample_info.csv' (see source url) and details on data generation can be found in Ghandi, M., et al (2019) (https://www.nature.com/articles/s41586-019-1186-3).
}
\keyword{datasets}
