% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ARID1A_KO_screen.R
\docType{data}
\name{sample_ARID1A_KO_screen}
\alias{sample_ARID1A_KO_screen}
\title{Sample GINI screen result for ARID1A KO pan cancer cell lines}
\format{
A data frame with 18,333 rows and 17 variables:
}
\usage{
sample_ARID1A_KO_screen
}
\description{
A sample data set containing the results of a screen conducted in cancer cell lines with
loss-of-function alterations in ARID1A (HomDel) generated using \code{GINI_screen()}. As reported in previous literature
ARID1A and ARID1B have synthetic lethal interaction (Helming et al.; doi: 10.1038/nm.3480; PMID:24562383),
thus gene with the highest interaction score is ARID1B.
}
\keyword{datasets}
