% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cell_lines.R
\name{select_cell_lines}
\alias{select_cell_lines}
\title{Select mutant groups based on input gene of interest}
\usage{
select_cell_lines(
  Input_gene = NULL,
  Input_disease = NULL,
  Input_disease_subtype = NULL
)
}
\arguments{
\item{Input_gene}{string Input Hugo Symbol}

\item{Input_disease}{string Cancer type selected from \code{list_available_cancer_types()}}

\item{Input_disease_subtype}{string Cancer subtype listed in \code{list_available_cancer_subtypes()}}
}
\value{
data frame containing a summary of mutations found in cell lines and their control and mutant group assignments.
}
\description{
\code{select_cell_lines()} assigns cancer cell lines to either \code{Control} groups or one of the following mutant groups: \code{HomDel},
\code{T-HetDel}, \code{HetDel}, \code{Amplified}, or \code{Others} (see below for more details).
}
\details{
Mutant groups in more detail:
\itemize{
\item Control` cell lines do not harbor any single nucleotide variations (SNVs) or insertions and deletions (InDels) with a neutral copy number (CN).
\item \code{HomDel} cell lines harbor one or more homozygous deleterious SNVs or have deep CN loss.
\item \code{T-HetDel} cell lines harbor two or more heterozygous deleterious SNVs/InDels with neutral or CN loss.
\item \code{HetDel} cell lines harbor one heterozygous deleterious SNV/InDel with neutral CN, or no SNV/InDel with CN loss.
\item \code{Amplified} cell lines harbor no SNVs/InDels with increased CN.
\item \code{Others} cell lines harbor deleterious SNVs with increased CN.
}
}
\examples{
\dontrun{
# Looking for TP53 mutants in all cancer cell lines (pan-cancer search)
select_cell_lines(Input_gene = "TP53")

# TP53 mutants only in SCLC subtypes
select_cell_lines(Input_gene = "TP53", Input_disease = "Lung Cancer", Input_disease_subtype = "Small Cell Lung Cancer (SCLC)")

# All cancer cell lines that are from Lung Cancers
select_cell_lines(Input_disease = "Lung Cancer")
}
}
