% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_mutant_groups.R
\name{select_mutant_groups}
\alias{select_mutant_groups}
\title{Select mutant groups based on input gene of interest}
\usage{
select_mutant_groups(Input_gene)
}
\arguments{
\item{Input_gene}{string Input Hugo Symbol}
}
\value{
data frame containing a summary of mutations found in cell lines and their control and mutant group assignments.
}
\description{
\code{select_mutant_groups()} assigns cancer cell lines to either \code{Control} groups or one of the following mutant groups: \code{HomDel},
\code{T-HetDel}, \code{HetDel}, \code{Amplified}, or \code{Others} (see below for more details).
}
\details{
Mutant groups in more detail:
\itemize{
\item Control` cell lines do not harbor any single nucleotide variations (SNVs) or insertions and deletions (InDels) with a neutral copy number (CN).
\item \code{HomDel} cell lines harbor one or more homozygous deleterious SNVs or have deep CN loss.
\item \code{T-HetDel} cell lines harbor two or more heterozygous deleterious SNVs/InDels with neutral or CN loss.
\item \code{HetDel} cell lines harbor one heterozygous deleterious SNV/InDel with neutral CN, or no SNV/InDel with CN loss.
\item \code{Amplified} cell lines harbor no SNVs/InDels with increased CN.
\item \code{Others} cell lines harbor deleterious SNVs with increased CN.
}
}
\examples{
A1CF_groups <- select_mutant_groups("A1CF")
}
