% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Append_perm_pval.R
\name{Append_perm_pval}
\alias{Append_perm_pval}
\title{Appends permutation adjusted p-value}
\usage{
Append_perm_pval(GI_screen_res = NULL, perms = NULL)
}
\arguments{
\item{GI_screen_res}{dataframe, Output from GI_screen(), Default: NULL}

\item{perms}{dataframe, Output from Pval_perm_adjust(), Default: NULL}
}
\value{
A dataframe with permutation adjusted p-values appeneded as a new column named Pval_perm_adj.
}
\description{
Appends permutation adjusted p-values to new column.
}
\details{
P-values are adjusted based on permuted random sampled background data.
}
\examples{
# Please see tutorial

}
