% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_prot.R
\name{extract_prot}
\alias{extract_prot}
\title{Extract protein expression data for given cell lines}
\usage{
extract_prot(input_samples = NULL, input_genes = NULL, data_dir = NULL)
}
\arguments{
\item{input_samples}{string A vector of DepMap_ID(s) must be provided, Default: NULL}

\item{input_genes}{string Optional Hugo Symbol(s) encoding proteins of interest, Default: NULL}

\item{data_dir}{string Path to GINIR_data}
}
\value{
Data frame containing protein expression for samples provided in the input.
If no genes were specified, the function will return a data frame of proteins profiled in DepMap
}
\description{
FUNCTION_DESCRIPTION
}
\details{
See also \code{extract_protein_rna} to extract proteomics profile data
}
\examples{
gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

if(!dir.exists(gretta_data_dir)){
  download_example_data(".")
}

extract_prot(
input_samples = c('ACH-000004', 'ACH-000146'), 
input_genes = c('CIC'),
data_dir = gretta_data_dir)

}
