% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rna_expr.R
\name{extract_rna_expr}
\alias{extract_rna_expr}
\title{Extract RNA expression data for given cell lines}
\usage{
extract_rna_expr(Input_samples = NULL, Input_genes = NULL)
}
\arguments{
\item{Input_samples}{string A vector of DepMap_ID(s) must be provided, Default: NULL}

\item{Input_genes}{string Optional Hugo Symbol(s), Default: NULL}
}
\value{
Data frame containing RNA expression (TPM) for sample provided in the input.
If no genes were specified, the function will return a data frame of all genes profiled in DepMap
}
\description{
FUNCTION_DESCRIPTION
}
\details{
See also \code{extract_protein_expr} to extract proteomics profile data
}
\examples{
\dontrun{
extract_rna_expr(Input_samples = c("ACH-001642","ACH-000688"), Input_genes = c("TP53","ARID1A"))
}
}
