% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_GeneNameID}
\alias{get_GeneNameID}
\alias{get_DepMapID}
\title{Get unique DepMap-compatible gene IDs}
\usage{
get_GeneNameID(gene_name, data_dir)

get_DepMapID(sample_name, data_dir)
}
\arguments{
\item{gene_name}{string containing a vector of either Hugo gene symbol or numeric NCBI ID}

\item{data_dir}{string Path to GINIR_data}

\item{sample_name}{string containing a vector of unique sample_id used in proteomics data or common cell line names}
}
\value{
string

string
}
\description{
\code{get_GeneNameID()} and \code{get_DepMapID()} provide tools for converting gene symbols/ids and
common cell line aliases to DepMap's unique gene name and cell line identifiers, respectively.
}
\section{Functions}{
\itemize{
\item \code{get_DepMapID()}: Get unique DepMap-compatible sample IDs

}}
\examples{
gretta_data_dir <- '/projects/marralab/ytakemon_prj/DepMap/GRETTA_data/22Q2/data'

get_GeneNameID('A1CF', data_dir = gretta_data_dir)

gretta_data_dir <- '/projects/marralab/ytakemon_prj/DepMap/GRETTA_data/22Q2/data'
gretta_output_dir <- '/projects/marralab/ytakemon_prj/DepMap/GRETTA_troubleshooting/'

get_DepMapID('JURKAT', data_dir = gretta_data_dir)

}
