% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_GeneNameID}
\alias{get_GeneNameID}
\alias{get_DepMapID}
\title{Get unique DepMap-compatible gene IDs}
\usage{
get_GeneNameID(GeneName, data_dir)

get_DepMapID(sample_name, data_dir)
}
\arguments{
\item{GeneName}{string containing a vector of either Hugo gene symbol or numeric NCBI ID}

\item{data_dir}{string Path to GINIR_data}

\item{sample_name}{string containing a vector of unique sample_id used in proteomics data or common cell line names}
}
\value{
string

string
}
\description{
\code{get_GeneNameID()} and \code{get_DepMapID()} provide tools for converting gene symbols/ids and
common cell line aliases to DepMap's unique gene name and cell line identifiers, respectively.
}
\section{Functions}{
\itemize{
\item \code{get_DepMapID}: Get unique DepMap-compatible sample IDs
}}

\examples{
\dontrun{
get_GeneNameID("A1CF", data_dir = "/path/to/DepMap_data/")
}
\dontrun{
get_DepMapID("JURKAT", data_dir = "/path/")
}
}
