% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cell_lines.R
\name{select_cell_lines}
\alias{select_cell_lines}
\title{Select mutant groups based on input gene of interest}
\usage{
select_cell_lines(
  input_gene = NULL,
  input_aa_change = NULL,
  input_disease = NULL,
  input_disease_subtype = NULL,
  data_dir = NULL
)
}
\arguments{
\item{input_gene}{string as Hugo Symbol}

\item{input_aa_change}{string Amino acid change (eg. "A387A"). input_gene must be specified}

\item{input_disease}{string Cancer type listed in \code{list_available_cancer_types()}}

\item{input_disease_subtype}{string Cancer subtype listed in \code{list_available_cancer_subtypes()}}

\item{data_dir}{string Path to GINIR_data}
}
\value{
Data frame containing a summary of mutations found in cell lines and their control and mutant group assignments.
}
\description{
\code{select_cell_lines()} assigns cancer cell lines to either \code{Control} groups or one of the following mutant groups: \code{HomDel},
\code{T-HetDel}, \code{HetDel}, \code{Amplified}, or \code{Others} (see details).
}
\details{
Mutant groups in more detail when only \code{input_gene} is defined:
\itemize{
\item \code{Control} cell lines do not harbor any single nucleotide variations (SNVs) or insertions and deletions (InDels) with a neutral copy number (CN).
\item \code{HomDel} cell lines harbor one or more homozygous deleterious SNVs or have deep CN loss.
\item \code{T-HetDel} cell lines harbor two or more heterozygous deleterious SNVs/InDels with neutral or CN loss.
\item \code{HetDel} cell lines harbor one heterozygous deleterious SNV/InDel with neutral CN, or no SNV/InDel with CN loss.
\item \code{Amplified} cell lines harbor no SNVs/InDels with increased CN.
\item \code{Others} cell lines harbor deleterious SNVs with increased CN.
}

If input_aa_change` is also defined:
\itemize{
\item \code{Control} cell lines do not harbor any single nucleotide variations (SNVs) or insertions and deletions (InDels) with a neutral copy number (CN).
\item \code{HomAlt} cell lines harbor a homozygous alteration for the specified mutation.
\item \code{HetAlt} cell lines harbor a heterozygous alteration for the specified mutation.
\item \verb{_CNneutral}, \verb{_CNamplified}, and \verb{_CNloss} define copy number (CN) status of the above mutation states.
\item \code{Others} cell lines that do not meet above criteria.
}
}
\examples{
gretta_data_dir <- './GRETTA_example/'
gretta_output_dir <- './GRETTA_example_output/'

if(!dir.exists(gretta_data_dir)){
  download_example_data(".")
}

select_cell_lines(input_gene = "ARID1A", 
input_disease = "Lung Cancer", 
data_dir = gretta_data_dir)

}
